Float left proof
The idea is to use [Hugo's markdown render hooks](https://gohugo.io/templates/render-hooks/) to customise how an image is displayed with markdown.

We co-opt the alt text to allow us to add custom class name to the image. This way we can choose to float left, or right or put the image in the middle of the page or whatever. If the alt text starts with `.` and has a single `|` character, take the text between the `.` and `|` and use it as a class for the paragraph.

Usually you'd also want to run some regex validation on the text to make sure it's a valid classname and prevent DOM injection, but since its just generating static content, there's not the same need for security.

`.pull-left` is the old bootstrap class intended to add `float:left;`

Looks like this template is using a very old version of bootstrap, which might not be well documented, and can be especially painful to deal with. Might be tempted to start from scratch if I find the time one day.